Vector for Numeric types.
Changed the manner in which the type Vector operates. Vectors are generally
associated with mathematical operations and as a result, the Vector has
now been made a specialization of AbstractList<E> with Numeric as the
type parameter.

For general object containment in a list type structure, please use the
TypeList class which has been added.

Corrected a bug where the upper bound was being excluded in Vector and
TypeList.

+++ src/main/java/net/sourceforge/cilib/cooperative/CooperativeEntity.java
@@ -26,8 +26,10 @@ import net.sourceforge.cilib.entity.Entity;
 import net.sourceforge.cilib.problem.Fitness;
 import net.sourceforge.cilib.problem.InferiorFitness;
 import net.sourceforge.cilib.problem.OptimisationProblem;
+import net.sourceforge.cilib.type.types.Numeric;
 import net.sourceforge.cilib.type.types.Type;
 import net.sourceforge.cilib.type.types.container.StructuredType;
+import net.sourceforge.cilib.type.types.container.TypeList;
 import net.sourceforge.cilib.type.types.container.Vector;
 
 /**
@@ -89,7 +91,7 @@ public class CooperativeEntity extends AbstractEntity {
         if(value instanceof Vector)
             context.append((Vector) value);
         else
-            context.append(value);
+            context.append((Numeric) value);
     }
 
     public void append(Entity entity) {
